Background notification implemented
but the checkWeather() function is dummy cause there is no global function for checking weather & reimplementing the weather factory every time doesn't seem a good option to me. will fix this in the next commit. my current approach is to check the weather every hour or two & push notification if it changes.